pyproject.toml
[project]
name = "science-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "arxiv>=2.2.0",
    "colorlog>=6.9.0",
    "fuzzyfinder>=2.3.0",
    "ipython>=9.5.0",
    "langchain-core>=0.3.75",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.7",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "langgraph-supervisor>=0.0.29",
    "markitdown[all]>=0.1.3",
    "mcp[cli]>=1.13.1",
    "pymupdf>=1.26.4",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "typing-extensions>=4.15.0",
]
